requirements.txt
gradio
requests
Pillow
google-api-python-client
oauth2client